refactor : context 코드 제거 zustand로 통일
