+++ services/admin/src/store/useEditVoteStore.ts
@@ -1,0 +1,14 @@
+import { create } from 'zustand';
+
+interface EditVoteState {
+  editingSurveyId: string | null;
+  setEditingSurveyId: (id: string | null) => void;
+}
+
+export const useEditVoteStore = create<EditVoteState>()((set) => ({
+  editingSurveyId: null,
+  setEditingSurveyId: (id) =>
+    set(() => ({
+      editingSurveyId: id,
+    })),
+}));
